Common classes for position fragments (OSM/Google)
