+++ app/src/googleplay/res/layout/fragment_send_position_google.xml
@@ -51,7 +51,7 @@
             android:layout_width="24dp"
             android:layout_height="42dp"
             android:layout_gravity="center_horizontal"
-            android:src="@drawable/ic_map_pin"/>
+            android:src="@drawable/ic_map_pin_google"/>
 
         <ImageView
             android:id="@+id/pin_x"
