+++ app/src/main/res/drawable-hdpi/ic_map_pin_google.png
